Matrix: split sdk into focused crypto/http/type modules
